requirements.txt
Pillow
pygame
cryptography
cython
pyinstaller